feat: Move config to src folder
